# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zawadi HR & Payroll API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Principal"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeRow"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetail"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Terminate employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
  /employees/{id}/portal-access:
    post:
      operationId: grantPortalAccess
      tags: [employees]
      summary: Grant portal access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalAccessInput"
      responses:
        "201":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalAccessResult"
...
components:
  schemas:
...
    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        orgSlug:
          type: ["string", "null"]
    Principal:
      type: object
      required: [userId, orgId, email, name, role, orgSlug, countryCode, currencyCode, mustChangePassword]
      properties:
        userId:
          type: integer
        orgId:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        employeeId:
          type: ["integer", "null"]
        mustChangePassword:
          type: boolean
        orgSlug:
          type: string
        countryCode:
          type: string
        currencyCode:
          type: string
    LoginResult:
      type: object
      required: [user, org]
      properties:
        user:
          type: object
          required: [id, name, email, role, mustChangePassword]
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
            role:
              type: string
            mustChangePassword:
              type: boolean
        org:
          type: object
          required: [slug, name, countryCode, currencyCode]
          properties:
            slug:
              type: string
            name:
              type: string
            countryCode:
              type: string
            currencyCode:
              type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    Employee:
      type: object
      required: [id, orgId, empNo, firstName, lastName, email, position, employmentType, status, basicSalary, hireDate]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        empNo:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: string
        nationalId:
          type: ["string", "null"]
        kraPin:
          type: ["string", "null"]
        nssfNo:
          type: ["string", "null"]
        shifNo:
          type: ["string", "null"]
        payMethod:
          type: string
        bankName:
          type: ["string", "null"]
        bankAccount:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        position:
          type: string
        employmentType:
          type: string
        residentStatus:
          type: string
        disabilityExemption:
          type: boolean
        basicSalary:
          type: integer
        houseAllowance:
          type: integer
        transportAllowance:
          type: integer
        otherAllowance:
          type: integer
        nonCashBenefit:
          type: integer
        insurancePremium:
          type: integer
        pensionEmployee:
          type: integer
        pensionEmployer:
          type: integer
        mortgageInterest:
          type: integer
        helbMonthly:
          type: integer
        saccoMonthly:
          type: integer
        hireDate:
          type: string
        terminationDate:
          type: ["string", "null"]
        status:
          type: string
        leaveBalance:
          type: integer
        createdAt:
          type: string
    EmployeeRow:
      type: object
      required: [employee]
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        department:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
    EmployeeDetail:
      type: object
      required: [employee]
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        department:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
        payslips:
          type: array
          items:
            type: object
            additionalProperties: true
        loans:
          type: array
          items:
            type: object
            additionalProperties: true
    EmployeeInput:
      type: object
      required: [firstName, lastName, position, employmentType, residentStatus, basicSalary, hireDate]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          type: string
        nationalId:
          type: string
        kraPin:
          type: string
        nssfNo:
          type: string
        shifNo:
          type: string
        payMethod:
          type: string
        bankName:
          type: string
        bankCode:
          type: string
        bankBranchCode:
          type: string
        bankAccount:
          type: string
        mpesaPhone:
          type: string
        departmentId:
          type: integer
        position:
          type: string
        employmentType:
          type: string
        residentStatus:
          type: string
        disabilityExemption:
          type: boolean
        basicSalary:
          type: number
        houseAllowance:
          type: number
        transportAllowance:
          type: number
        otherAllowance:
          type: number
        nonCashBenefit:
          type: number
        insurancePremium:
          type: number
        pensionEmployee:
          type: number
        pensionEmployer:
          type: number
        mortgageInterest:
          type: number
        helbMonthly:
          type: number
        saccoMonthly:
          type: number
        hireDate:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        departmentId:
          type: integer
        employmentType:
          type: string
        residentStatus:
          type: string
        disabilityExemption:
          type: boolean
        basicSalary:
          type: number
        houseAllowance:
          type: number
        transportAllowance:
          type: number
        otherAllowance:
          type: number
        nonCashBenefit:
          type: number
        insurancePremium:
          type: number
        pensionEmployee:
          type: number
        pensionEmployer:
          type: number
        mortgageInterest:
          type: number
        helbMonthly:
          type: number
        saccoMonthly:
          type: number
        kraPin:
          type: string
        nssfNo:
          type: string
        shifNo:
          type: string
        bankName:
          type: string
        bankAccount:
          type: string
        payMethod:
          type: string
        status:
          type: string
    PortalAccessInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    PortalAccessResult:
      type: object
      required: [tempPassword]
      properties:
        tempPassword:
          type: string
...